Address PR review comments: fix saveIfDirty race, 64-bit cache hash, enforce cache cap, long-word hyphenation fallback, snprintf format fix
- saveIfDirty(): Only clear dirty_ after successful saveToFile()
- Word-width cache: Upgraded to 64-bit FNV-1a hash to reduce collision risk
- Word-width cache: Enforce WORD_WIDTH_CACHE_MAX cap (skip insertion when full)
- Hyphenation: Fall back to std::string for prefixes > 126 bytes (long URLs/identifiers)
- setViewportDimensions: Cast uint16_t to unsigned int for %u format specifier

Agent-Logs-Url: https://github.com/m-zaw/cpr-vcodex/sessions/d2ad12a2-3182-48d5-81e2-ae7841d17caa

+++ lib/Epub/Epub/Section.h
@@ -46,7 +46,8 @@ class Section {
   // other orientation's cache — the user can switch back without a full rebuild.
   void setViewportDimensions(uint16_t viewportWidth, uint16_t viewportHeight) {
     char buf[16];
-    snprintf(buf, sizeof(buf), "_%ux%u", viewportWidth, viewportHeight);
+    snprintf(buf, sizeof(buf), "_%ux%u", static_cast<unsigned int>(viewportWidth),
+             static_cast<unsigned int>(viewportHeight));
     filePath = epub->getCachePath() + "/sections/" + std::to_string(spineIndex) + buf + ".bin";
     pageLut.clear();  // New file path, invalidate any cached LUT
   }
